added if-checks for all body-parameters in server.js. it will now respond with "error" if any parameter is missing/undefined

+++ server/src/server.js
@@ -52,14 +52,18 @@ let publicKey: string;
 
 let privateKey = (publicKey = "shhhhhverysecret");
 
-// middleware-function
+// middleware-functiona
 
 
 app.use("/api", (req, res, next) => {
+
+    if(req.headers['x-access-token'] === undefined) return res.json({error: "missing access token header"});
+
     var token = req.headers["x-access-token"];
 
     if(token === MASTER_TOKEN){
         req.email = req.body.email;
+        req.userId = req.body.userId;
         next();
     } else {
         jwt.verify(token, publicKey, (err, decoded) => {
@@ -71,12 +75,12 @@ app.use("/api", (req, res, next) => {
                 userDao.getUser(req.body.userId, (status, data) => {
                     console.log(decoded);
                     console.log(data);
-                    console.log("OSOADUSDIOSABDOUASDBOAUSDBUSODA");
                     if(data.length === 0){
                         res.status(401);
                         res.json({error : "Not authorized -> Invalid Token"});
                     } else if(data[0].email === decoded.email) {
                         req.email = decoded.email;
+                        req.userId = data[0].userId;
                         next();
                     } else {
                         res.status(401);
@@ -119,9 +123,10 @@ app.get("/api/users", (req, res) => {
 });
 
 //Delete rider
-app.delete('/api/event/:event_id/rider', (req, res) => {
+app.delete('/api/performance/:performance_id/rider', (req, res) => {
     console.log("Fikk DELETE-request fra klienten");
-    eventDao.deleteRider(req.body, (status, data) => {
+    if(req.body.name === undefined) return res.json({error : "bad request"});
+    eventDao.deleteRider({performanceId : req.params.performance_id, name : req.body.name}, (status, data) => {
         res.status(status);
         let token =  thisFunctionCreatesNewToken(req.email);
         res.json({data, jwt: token});
@@ -131,8 +136,9 @@ app.delete('/api/event/:event_id/rider', (req, res) => {
 // DELETE TICKET
 
 app.delete("/api/event/:event_id/ticket", (req, res) => {
+    if(req.body.name === undefined) return res.json({error : "bad request"});
     console.log("Fikk Delete-request fra klient");
-    eventDao.deleteTicket(req.body, (status, data) => {
+    eventDao.deleteTicket({eventId : req.params.event_id, name : req.body.name}, (status, data) => {
         res.status(status);
         let token =  thisFunctionCreatesNewToken(req.email);
         res.json({data, jwt: token});
@@ -142,8 +148,8 @@ app.delete("/api/event/:event_id/ticket", (req, res) => {
 
 // DELETE PERFORMANCE
 
-app.delete("/api/event/:event_id/performance", (req, res) => {
-    eventDao.deletePerformance(req.body, (status, data) => {
+app.delete("/api/performance/:performance_id", (req, res) => {
+    eventDao.deletePerformance(req.params.performance_id, (status, data) => {
         res.status(status);
         let token =  thisFunctionCreatesNewToken(req.email);
         res.json({data, jwt: token});
@@ -152,8 +158,9 @@ app.delete("/api/event/:event_id/performance", (req, res) => {
 
 // UPDATE PERFORMANCE
 
-app.put("/api/event/:event_id/performance", (req, res) => {
-    eventDao.updatePerformance(req.body, (status, data) => {
+app.put("/api/performance/:performance_id", (req, res) => {
+    if(req.body.startTime === undefined || req.body.endTime === undefined || req.body.contract === undefined) return res.json({error : "bad request"});
+    eventDao.updatePerformance({startTime : req.body.startTime, endTime : req.body.endTime, contract : req.body.contract, performanceId : req.params.performance_id}, (status, data) => {
         res.status(status);
         let token =  thisFunctionCreatesNewToken(req.email);
         res.json({data, jwt: token});
@@ -164,6 +171,8 @@ app.put("/api/event/:event_id/performance", (req, res) => {
 app.delete("/api/user/:user_id", (req, res) => {
 
     console.log("Fikk DELETE-request fra klienten");
+
+    if(req.body.password === undefined) return res.json({error : "request missing password"});
 
     userDao.getPassword(req.params.user_id, (status, data) => {
 
@@ -191,8 +200,13 @@ app.delete("/api/user/:user_id", (req, res) => {
 
 //Update a user
 app.put("/api/user/:user_id", (req, res) => {
-    console.log("Fikk PUT-requesr fra klienten");
-    userDao.updateUser(req.body, (status, data) => {
+    console.log("Fikk PUT-request fra klienten");
+    if(req.body.username === undefined || req.body.username === null) return res.json({error : "request missing username"});
+    if(req.body.email === undefined || req.body.email === null) return res.json({error : "request missing email"});
+    if(req.body.firstName === undefined) return res.json({error: "request missing first-name"});
+    if(req.body.lastName === undefined) return res.json({error : "requst misssing last-name"}); 
+    let params = {username : req.body.username, email : req.body.email, phone : req.body.phone, firstName : req.body.firstName, lastName : req.body.lastName, userId : req.params.userId}
+    userDao.updateUser(req.params, (status, data) => {
         res.status(status);
         let token =  thisFunctionCreatesNewToken(req.email);
         res.json({data, jwt: token});
@@ -203,6 +217,10 @@ app.put("/api/user/:user_id", (req, res) => {
 app.post("/user", (req, res) => {
 
     console.log("Fikk POST-request fra klienten");
+
+    if(req.body.username === undefined || req.body.username === null ) return res.json({error : "request missing username"});
+    if(req.body.password === undefined) return res.json({error : "request missing password"});
+    if(req.body.email === undefined || req.body.email === null ) return res.json({error : "request missing email"}); 
 
     let user = req.body;
 
@@ -243,6 +261,9 @@ app.get("/api/event/:event_id", (req, res) => {
 //post an event
 app.post("/api/event", (req, res) => {
     console.log("Fikk POST-request fra klienten");
+    if(req.body.name === undefined || req.body.name === null) return res.json({error : "request missing event-name"});
+    if(req.body.userId === undefined || req.body.userId === null) return res.json({error : "request missing event-host user Id"});
+
     eventDao.createEvent(req.body, (status, data) => {
         console.log("sdasa",data);
         res.status(status);
@@ -254,6 +275,10 @@ app.post("/api/event", (req, res) => {
 //post a ticket
 app.post("/api/event/:event_id/ticket", (req, res) => {
     console.log("Fikk POST-request fra klienten");
+    if(req.body.name === null || req.body.name === undefined) return res.json({error : "post-request missing ticket name"});
+    if(req.body.eventId === null || req.body.eventId === undefined) return res.json({error : "post-request : ticket missing event"});
+    if(req.body.price === null || req.body.price === undefined) return res.json({error : "post-request ticket missing ticket-price"});
+    if(req.body.amount === null || req.body.amount === undefined) return res.json({error : "post request missing number of tickets"});
     eventDao.createTicket(req.body, (status, data) => {
         res.status(status);
         let token =  thisFunctionCreatesNewToken(req.email);
@@ -264,6 +289,9 @@ app.post("/api/event/:event_id/ticket", (req, res) => {
 //put ticket
 app.put("/api/event/:event_id/ticket", (req, res) => {
     console.log("Fikk PUT-request fra klienten");
+    if(req.body.price === undefined || req.body.price === null) return res.json({error : "request missing ticket - price"});
+    if(req.body.amount == undefined) return res.json({error : "request missing number of tickets"});
+    if(req.body.name == undefined) return res.json({error : "request missing ticket-name"});
     eventDao.updateTicket(req.body, (status, data) => {
         res.status(status);
         let token =  thisFunctionCreatesNewToken(req.email);
@@ -272,9 +300,14 @@ app.put("/api/event/:event_id/ticket", (req, res) => {
 });
 
 //put rider
-app.put("/api/event/:event_id/rider", (req, res) => {
+app.put("/api/performance/:performance_id/rider", (req, res) => {
     console.log("Fikk PUT-request fra klienten");
-    eventDao.updateRider(req.body, (status, data) => {
+
+    if(req.body.name == undefined) return res.json({error : "bad request : missing name"});
+    if(req.body.amount == undefined) return res.json({error : "bad request : missing amount of tickets"});
+    if(req.body.oldName == undefined) return res.json({error : "bad request : missing old name - parameter"});
+
+    eventDao.updateRider({name : req.body.name, amount : req.body.amount, performanceId : req.params.performance_id, oldName : req.body.oldName}, (status, data) => {
         res.status(status);
         let token =  thisFunctionCreatesNewToken(req.email);
         res.json({data, jwt: token});
@@ -282,9 +315,13 @@ app.put("/api/event/:event_id/rider", (req, res) => {
 });
 
 //post a performance
-app.post("api/event/:event_id/user", (req, res) => {
+app.post("api/event/:event_id/performance", (req, res) => {
     console.log("Fikk POST-request fra klienten");
-    eventDao.createPerformance(req.body, (status, data) => {
+    if(req.body.artistId == undefined) return res.json({error : "bad request : missing artist parameter"});
+    if(req.body.startTime == undefined) return res.json({error : "bad requst : missing startTime parameter"});
+    if(req.body.endTime == undefined) return res.json({error : "bad request : missing endTime parameter"});
+    if(req.body.contract == undefined) req.body.contract = "";
+    eventDao.createPerformance({artistId: req.body.artistId, eventId: req.params.event_id, startTime : req.body.startTime, endTime : req.body.endTime, contract: req.body.contract}, (status, data) => {
         res.status(status);
         let token =  thisFunctionCreatesNewToken(req.email);
         res.json({data, jwt: token});
@@ -292,9 +329,11 @@ app.post("api/event/:event_id/user", (req, res) => {
 });
 
 //post a rider
-app.post("/api/event/:event_id/rider", (req, res) => {
+app.post("/api/performance/:performance_id/rider", (req, res) => {
     console.log("Fikk POST-request fra klienten");
-    eventDao.createRider(req.body, (status, data) => {
+    if(req.body.amount == undefined) return res.json({error : "bad request : missing amount parameter"});
+
+    eventDao.createRider({performanceId : req.params.performance_id, name : req.params.name, amount : req.body.amount}, (status, data) => {
         res.status(status);
         let token =  thisFunctionCreatesNewToken(req.email);
         res.json({data, jwt: token});
@@ -304,7 +343,15 @@ app.post("/api/event/:event_id/rider", (req, res) => {
 //Update an event
 app.put("/api/event/:event_id", (req, res) => {
     console.log("Fikk PUT-request fra klienten");
-    eventDao.updateEvent(req.body, (status, data) => {
+
+    if(req.body.eventName == undefined) return res.json({error : "bad request : missing eventName parameter"});
+    if(req.body.startTime == undefined) return res.json({error : "bad request : missing startTime parameter"});
+    if(req.body.hostId == undefined) return res.json({error : "bad request : missing hostId parameter"});
+    if(req.body.active == undefined) req.body.active = 1;
+    if(req.body.location == undefined) return res.json({error : "bad request : mssing location parameter"});
+
+    eventDao.updateEvent({eventName : req.body.eventName, hostId : req.body.hostId, active : req.body.active, 
+    location : req.body.location , startTime : req.body.startTime, endTime: req.body.endTime, eventId : req.params.event_id}, (status, data) => {
         res.status(status);
         let token =  thisFunctionCreatesNewToken(req.email);
         res.json({data, jwt: token});
@@ -314,11 +361,14 @@ app.put("/api/event/:event_id", (req, res) => {
 //Delete an event
 app.delete("/api/event/:event_id", (req, res) => {
     console.log("Fikk DELETE-request fra klienten");
-    eventDao.deleteEvent(req.params.event_id, (status, data) => {
-
-        res.status(status);
-        let token =  thisFunctionCreatesNewToken(req.email);
-        res.json({data, jwt: token});
+    eventDao.getEvent(req.params.event_id, (status, data) => {
+        if(data[0].user_id === req.userId){    
+            eventDao.deleteEvent(req.params.event_id, (status, data) => {
+                res.status(status);
+                let token =  thisFunctionCreatesNewToken(req.email);
+                res.json({data, jwt: token});
+            });
+        } else res.json({error : "not authorized"});
     });
 });
 
@@ -450,6 +500,9 @@ app.put("/user/:usermail", (req, res) => {
 // Handles login and returns JWT-token as JSON
 app.post("/login", (req, res) => {
 
+    if(req.body.email == undefined) return res.json({error : "bad request : missing email parameter"});
+    if(req.body.password == undefined) return res.json({error : " bad request : mssing password parameter"})
+
     loginOk(req.body.email, req.body.password).then(data => {
         let log: number = data;
         console.log("User ID", log);
@@ -467,7 +520,7 @@ app.post("/login", (req, res) => {
         }
     });
 
-    async function loginOk(email: string, pw: string): Promise {
+    async function loginOk(email: string, pw: string): Promise<any> {
         return new Promise(resolve => {
             userDao.getPassword(email, (status, data) => {
                 let log = -1;
